+++ plugin/src/main/java/net/aufdemrand/denizen/objects/dCuboid.java
@@ -1,10 +1,11 @@
 package net.aufdemrand.denizen.objects;
 
 import net.aufdemrand.denizen.Settings;
+import net.aufdemrand.denizen.nms.NMSHandler;
 import net.aufdemrand.denizen.nms.interfaces.BlockData;
+import net.aufdemrand.denizen.nms.interfaces.BlockHelper;
 import net.aufdemrand.denizen.objects.notable.NotableManager;
 import net.aufdemrand.denizen.utilities.Utilities;
-import net.aufdemrand.denizen.utilities.blocks.SafeBlock;
 import net.aufdemrand.denizen.utilities.debugging.dB;
 import net.aufdemrand.denizen.utilities.depends.Depends;
 import net.aufdemrand.denizencore.objects.*;
@@ -599,6 +600,7 @@ public class dCuboid implements dObject, Cloneable, Notable, Adjustable {
         dList list = new dList();
         int index = 0;
 
+        BlockHelper blockHelper = NMSHandler.getInstance().getBlockHelper();
         for (LocationPair pair : pairs) {
 
             dLocation loc_1 = pair.low;
@@ -612,8 +614,8 @@ public class dCuboid implements dObject, Cloneable, Notable, Adjustable {
                         loc = new dLocation(loc_1.clone()
                                 .add(x, y, z));
 
-                        if (SafeBlock.blockIsSafe(loc.getBlock().getType())
-                                && SafeBlock.blockIsSafe(loc.clone().add(0, 1, 0).getBlock().getType())
+                        if (blockHelper.isSafeBlock(loc.getBlock().getType())
+                                && blockHelper.isSafeBlock(loc.clone().add(0, 1, 0).getBlock().getType())
                                 && loc.clone().add(0, -1, 0).getBlock().getType().isSolid()
                                 && matchesMaterialList(loc.clone().add(0, -1, 0), mats)) {
                             // Get the center of the block, so the entity won't suffocate
