+++ v1_10_R1/src/main/java/net/aufdemrand/denizen/nms/helpers/BlockHelper_v1_10_R1.java
@@ -8,6 +8,7 @@ import net.aufdemrand.denizen.nms.interfaces.BlockData;
 import net.aufdemrand.denizen.nms.interfaces.BlockHelper;
 import net.aufdemrand.denizen.nms.util.PlayerProfile;
 import net.minecraft.server.v1_10_R1.TileEntitySkull;
+import org.bukkit.Material;
 import org.bukkit.SkullType;
 import org.bukkit.block.Block;
 import org.bukkit.block.Skull;
@@ -55,5 +56,44 @@ public class BlockHelper_v1_10_R1 implements BlockHelper {
     @Override
     public BlockData getBlockData(String compressedString) {
         return BlockData_v1_10_R1.fromCompressedString(compressedString);
+    }
+
+    @Override
+    public boolean isSafeBlock(Material material) {
+        // Quick util function to decide whether
+        // A block is 'safe' (Can be spawned inside of) - air, tallgrass, etc.
+        // Credit to Mythan for compiling the initial list
+        switch (material) {
+            case LEVER:
+            case WOOD_BUTTON:
+            case STONE_BUTTON:
+            case REDSTONE_WIRE:
+            case SAPLING:
+            case SIGN_POST:
+            case WALL_SIGN:
+            case SNOW:
+            case TORCH:
+            case DETECTOR_RAIL:
+            case ACTIVATOR_RAIL:
+            case RAILS:
+            case POWERED_RAIL:
+            case NETHER_WARTS:
+            case NETHER_STALK:
+            case VINE:
+            case SUGAR_CANE_BLOCK:
+            case CROPS:
+            case LONG_GRASS:
+            case RED_MUSHROOM:
+            case BROWN_MUSHROOM:
+            case DEAD_BUSH:
+            case REDSTONE_TORCH_OFF:
+            case REDSTONE_TORCH_ON:
+            case AIR:
+            case YELLOW_FLOWER:
+            case RED_ROSE:
+                return true;
+            default:
+                return false;
+        }
     }
 }
