+++ v1_10_R1/src/main/java/net/aufdemrand/denizen/nms/impl/packets/handlers/DenizenPacketListener_v1_10_R1.java
@@ -1,5 +1,6 @@
 package net.aufdemrand.denizen.nms.impl.packets.handlers;
 
+import net.aufdemrand.denizen.nms.NMSHandler;
 import net.aufdemrand.denizen.nms.impl.packets.PacketInResourcePackStatus_v1_10_R1;
 import net.aufdemrand.denizen.nms.impl.packets.PacketInSteerVehicle_v1_10_R1;
 import net.aufdemrand.denizen.nms.interfaces.packets.PacketHandler;
@@ -8,25 +9,23 @@ import net.minecraft.server.v1_10_R1.NetworkManager;
 import net.minecraft.server.v1_10_R1.Packet;
 import net.minecraft.server.v1_10_R1.PacketPlayInResourcePackStatus;
 import net.minecraft.server.v1_10_R1.PacketPlayInSteerVehicle;
+import org.bukkit.Bukkit;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.EventPriority;
 import org.bukkit.event.Listener;
 import org.bukkit.event.player.PlayerJoinEvent;
-import org.bukkit.plugin.java.JavaPlugin;
 
 public class DenizenPacketListener_v1_10_R1 extends AbstractListenerPlayIn_v1_10_R1 {
 
-    private static JavaPlugin plugin;
     private static PacketHandler packetHandler;
 
     public DenizenPacketListener_v1_10_R1(NetworkManager networkManager, EntityPlayer entityPlayer) {
         super(networkManager, entityPlayer, entityPlayer.playerConnection);
     }
 
-    public static void enable(JavaPlugin javaPlugin, PacketHandler handler) {
-        plugin = javaPlugin;
+    public static void enable(PacketHandler handler) {
         packetHandler = handler;
-        javaPlugin.getServer().getPluginManager().registerEvents(new PlayerEventListener(), javaPlugin);
+        Bukkit.getServer().getPluginManager().registerEvents(new PlayerEventListener(), NMSHandler.getJavaPlugin());
     }
 
     @Override
@@ -51,7 +50,7 @@ public class DenizenPacketListener_v1_10_R1 extends AbstractListenerPlayIn_v1_10
     public static class PlayerEventListener implements Listener {
         @EventHandler(priority = EventPriority.LOWEST)
         public void onPlayerJoin(PlayerJoinEvent event) {
-            DenizenNetworkManager_v1_10_R1.setNetworkManager(event.getPlayer(), plugin, packetHandler);
+            DenizenNetworkManager_v1_10_R1.setNetworkManager(event.getPlayer(), packetHandler);
         }
     }
 }
